Qwen_llama/requirements.txt
requests==2.32.5
duckdb==1.3.2
python-dotenv==1.2.2
reportlab==4.4.1
